+++ GUIManagerPatch.cs
@@ -1,9 +1,0 @@
-﻿using System;
-using HarmonyLib;
-
-public class Class1
-{
-	public Class1()
-	{
-	}
-}