+++ src/BuddyClimb/Patches/CharacterInteractiblePatch.cs
@@ -1,0 +1,162 @@
+using HarmonyLib;
+using Photon.Pun;
+
+namespace BuddyClimb.Patches;
+
+[HarmonyPatch(typeof(CharacterInteractible))]
+internal static class CharacterInteractiblePatch
+{
+    [HarmonyPatch(nameof(CharacterInteractible.GetInteractionText))]
+    [HarmonyPostfix]
+    private static void GetInteractionTextPatch(ref string __result, CharacterInteractible __instance)
+    {
+        if (__result != string.Empty)
+        {
+            return;
+        }
+
+        if (CanBeClimbed(__instance.character))
+        {
+            __result = "爬上去!";
+        }
+    }
+
+    [HarmonyPatch(nameof(CharacterInteractible.Interact))]
+    [HarmonyPostfix]
+    private static void InteractPatch(CharacterInteractible __instance, ref Character interactor)
+    {
+        if (__instance.CarriedByLocalCharacter() || (!__instance.IsCannibal() && __instance.CanBeCarried()))
+        {
+            return;
+        }
+
+        if (CanBeClimbed(__instance.character) && CanClimb(interactor))
+        {
+            interactor.photonView.RPC("RPCA_PassOut", __instance.character.player.photonView.Owner);
+            __instance.character.refs.carriying.StartCarry(interactor);
+        }
+    }
+
+    [HarmonyPatch(nameof(CharacterInteractible.IsPrimaryInteractible))]
+    [HarmonyPostfix]
+    private static void IsPrimaryInteractiblePatch(CharacterInteractible __instance, ref bool __result, ref Character interactor)
+    {
+        if (__result)
+        {
+            return;
+        }
+
+        if (CanBeClimbed(__instance.character) && CanClimb(interactor))
+        {
+            __result = true;
+        }
+    }
+
+    [HarmonyPatch(typeof(CharacterCarrying), nameof(CharacterCarrying.RPCA_Drop))]
+    [HarmonyPrefix]
+    private static bool RPCA_DropPrefix(PhotonView targetView)
+    {
+        if (targetView == null || !targetView.IsMine)
+        {
+            return true;
+        }
+
+        Character character = targetView.GetComponent<Character>();
+        Character? carrier = character?.data.carrier;
+        if (character != null && carrier != null && !character.data.fullyPassedOut)
+        {
+            targetView.RPC("RPCA_UnPassOut", carrier.photonView.Owner);
+        }
+
+        return true;
+    }
+
+    private static bool CanBeClimbed(Character character)
+    {
+        if (character.IsLocal)
+        {
+            return false;
+        }
+
+        if (character.data.dead)
+        {
+            return false;
+        }
+
+        if (character.player.backpackSlot.hasBackpack)
+        {
+            return false;
+        }
+
+        if (IsCharacterDoingIllegalCarryActions(character))
+        {
+            return false;
+        }
+
+        if (character.data.carriedPlayer)
+        {
+            return false;
+        }
+
+        if (character.data.carrier)
+        {
+            return false;
+        }
+
+        if (character.data.currentItem && character.data.currentItem.canUseOnFriend)
+        {
+            return false;
+        }
+
+        if (character.refs.customization.isCannibalizable)
+        {
+            return false;
+        }
+
+        return true;
+    }
+
+    private static bool CanClimb(Character character)
+    {
+        if (character.refs.interactible.CanBeCarried())
+        {
+            return true;
+        }
+
+        if (!character.IsLocal)
+        {
+            return false;
+        }
+
+        if (character.data.dead)
+        {
+            return false;
+        }
+
+        if (character.data.currentItem)
+        {
+            return false;
+        }
+
+        if (character.data.carrier)
+        {
+            return false;
+        }
+
+        if (character.refs.customization.isCannibalizable)
+        {
+            return false;
+        }
+
+        return true;
+    }
+
+    private static bool IsCharacterDoingIllegalCarryActions(Character character)
+    {
+        return character.data.isSprinting
+            || character.data.isJumping
+            || character.data.isClimbingAnything
+            || character.data.isCrouching
+            || character.data.isReaching;
+    }
+}